Create get-dinopass.psm1
Generates a password using the dinopass.com API with either the strong or simple switch. It will also display the password that's been retrieved and copy it to the clipboard. This allows you to choose a different password if it generates something uncomfortable, like "l@rgebLob23" which could yield the impression that you are insulting the end user.